:art: #20 - Create change user cover component.
