+++ src/components/user/common/UserCoverButton.tsx
@@ -4,32 +4,13 @@ import { FaRegEdit } from "react-icons/fa";
 import { updateUser, uploadImage } from "../../../api";
 import { UserType } from "../../../types";
 import ChangePicture from "../user-modals/ChangePicture";
+import ChangeUserCover from "./ChangeUserCover";
 import UserCover from "./UserCover";
 
 export default function UserCoverButton({ user }: { user: UserType }) {
   const [showCoverChangeModal, setShowCoverChangeModal] = useState<boolean>(
     false
   );
-
-  const toast = useToast();
-
-  const onDrop = useCallback(async (acceptedFiles: any) => {
-    const data = new FormData();
-    data.append("photo", acceptedFiles[0]);
-    const response = await uploadImage(data);
-    if (response.statusText === "OK") {
-      await updateUser({ coverImage: response.data });
-      window.location.reload();
-    } else {
-      toast({
-        title: "Can not upload image",
-        status: "error",
-        duration: 4000,
-        isClosable: true,
-      });
-    }
-  }, []);
-
   return (
     <Box>
       <UserCover user={user} />
@@ -52,13 +33,10 @@ export default function UserCoverButton({ user }: { user: UserType }) {
       >
         <FaRegEdit />
       </Button>
-      {showCoverChangeModal && (
-        <ChangePicture
-          open={showCoverChangeModal}
-          onClose={() => setShowCoverChangeModal(false)}
-          onDrop={onDrop}
-        />
-      )}
+      <ChangeUserCover
+        showCoverChangeModal={showCoverChangeModal}
+        setShowCoverChangeModal={setShowCoverChangeModal}
+      />
     </Box>
   );
 }
